% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_test.R
\name{BT}
\alias{BT}
\title{function to calculate sequence of test statistics by Busetti and Taylor (2004). For internal use only.}
\usage{
BT(x, trend, tau)
}
\description{
function to calculate sequence of test statistics by Busetti and Taylor (2004). For internal use only.
}
\keyword{internal}
