% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum_test.R
\name{cusum}
\alias{cusum}
\title{function to calculate sequence of cusum test statistics by Leybourne, Taylor, and Kim (2006). For internal use only}
\usage{
cusum(x, trend, m = 0, tau)
}
\description{
function to calculate sequence of cusum test statistics by Leybourne, Taylor, and Kim (2006). For internal use only
}
\keyword{internal}
