% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CritValues_Persistence.R
\name{CV}
\alias{CV}
\title{function to generate critical values. For internal use only.}
\usage{
CV(x, statistic, trend, type, m = 0, M, d = 0, tau, serial = FALSE,
  lmax = 0)
}
\description{
function to generate critical values. For internal use only.
}
\keyword{internal}
