% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_shift.R
\docType{data}
\name{d_vec}
\alias{d_vec}
\title{function to extract critical values. For internal use only.}
\format{An object of class \code{numeric} of length 11.}
\usage{
d_vec
}
\description{
function to extract critical values. For internal use only.
}
\keyword{internal}
