% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LKSN_test.R
\name{ers_test}
\alias{ers_test}
\title{Unit root test by Elliot et al. (1996). For internal use only}
\usage{
ers_test(y, trend, lag.max, T)
}
\description{
Unit root test by Elliot et al. (1996). For internal use only
}
\keyword{internal}
