% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_longrun.R
\name{fb_longrun}
\alias{fb_longrun}
\title{function to estimate the fixed-b long-run variance. For internal use only.}
\usage{
fb_longrun(tseries, m, type = "Bartlett")
}
\description{
function to estimate the fixed-b long-run variance. For internal use only.
}
\keyword{internal}
