% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixbsupw.R
\name{fixbsupw}
\alias{fixbsupw}
\title{Fixed-b sup Wald test for a single change in the mean of a long-memory time series.}
\usage{
fixbsupw(tseries, d, bandw = 0.1, tau = 0.15)
}
\arguments{
\item{tseries}{the univariate numeric time series to be investigated.}

\item{d}{the long-memory parameter.}

\item{bandw}{the bandwidth parameter for the long-run variance estimator. It can take values in the range \code{bandw=[0.05,0.1,0.2]}. Default is
\code{bandw=0.1}, which is suggested by Iacone, Leybourne and Taylor (2014).}

\item{tau}{defines the search area, which is \code{[tau,1-tau]}. Default is \code{tau=0.15} as suggested by Andrews (1993).}
}
\value{
Returns the test statistic and the corresponding critical values of the test.
}
\description{
This function performs a sup-Wald test on a change-in-mean, which is standardized by a non-parametric kernel-based long-run variance estimator.
Therefore, the test is robust under long-memory.
The function returns the test statistic as well as critical values.
}
\details{
Note that the critical values are generated for \code{tau=0.15} using the Bartlett kernel.
}
\examples{
library(fracdiff)
library(LongMemoryTS)

n        <- 500
d        <- 0.2
tseries  <- fracdiff.sim(n=n,d=d)$series
d_est    <- local.W(tseries, m=floor(1+n^0.65))$d

changep  <- c(rep(0,n/2),rep(1,n/2))
tseries2 <- tseries+changep
d_est2   <- local.W(tseries2, m=floor(1+n^0.65))$d

fixbsupw(tseries,d=d_est)
fixbsupw(tseries2,d=d_est2)
}
\references{
Iacone, F. and Leybourne, S. J. and Taylor, R. A. M. (2014): A fixed-b Test for a Break in Level at an unknown Time under Fractional Integration. Journal of Time Series Analysis, 35, pp. 40-54.

Andrews, D. W. K. (1993): Tests for Parameter Instability and Structural Change With Unknown Change Point. Econometrica, 61, pp. 821-856.
}
\author{
Kai Wenger
}
