% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CritValues_Persistence.R
\name{getCV}
\alias{getCV}
\title{function to extract critical values. For internal use only.}
\usage{
getCV()
}
\description{
function to extract critical values. For internal use only.
}
\keyword{internal}
