% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_test.R
\name{HLT}
\alias{HLT}
\title{function to calculate unit root test to adjust test statistic as suggested by Harvey, Leybourne, and Taylor (2006).}
\usage{
HLT(x, trend, z)
}
\description{
function to calculate unit root test to adjust test statistic as suggested by Harvey, Leybourne, and Taylor (2006).
}
\keyword{internal}
