% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_test.R
\name{HLTmin}
\alias{HLTmin}
\title{function to calculate sequence of minimum test statistics by Harvey, Leybourne, and Taylor (2006). For internal use only}
\usage{
HLTmin(x, trend, z, tau)
}
\description{
function to calculate sequence of minimum test statistics by Harvey, Leybourne, and Taylor (2006). For internal use only
}
\keyword{internal}
