% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBI_test.R
\name{LBI}
\alias{LBI}
\title{function to calculate sequence of cusum test statistics by Leybourne, Taylor, and Kim (2006). For internal use only}
\usage{
LBI(x, trend, m, tau)
}
\description{
function to calculate sequence of cusum test statistics by Leybourne, Taylor, and Kim (2006). For internal use only
}
\keyword{internal}
