% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LKSN_test.R
\name{LKSN}
\alias{LKSN}
\title{function to calculate sequence of LKSN test statistics. For internal use only}
\usage{
LKSN(x, trend, tau, lmax)
}
\description{
function to calculate sequence of LKSN test statistics. For internal use only
}
\keyword{internal}
