% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_test.R
\name{LT}
\alias{LT}
\title{function to calculate sequence of test statistics by Leybourne and Taylor (2004). For internal use only.}
\usage{
LT(x, trend, m, tau)
}
\description{
function to calculate sequence of test statistics by Leybourne and Taylor (2004). For internal use only.
}
\keyword{internal}
