% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR_test.R
\name{MR}
\alias{MR}
\title{function to calculate sequence of test statistics by Martins and Rodrigues (2014). For internal use only}
\usage{
MR(x, trend, serial, tau)
}
\description{
function to calculate sequence of test statistics by Martins and Rodrigues (2014). For internal use only
}
\keyword{internal}
