% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snwilcoxon.R
\name{snwilcoxon}
\alias{snwilcoxon}
\title{Self-normalized Wilcoxon test for a single change in the mean of a long-memory time series.}
\usage{
snwilcoxon(tseries, d, tau = 0.15)
}
\arguments{
\item{tseries}{the univariate numeric time series to be investigated.}

\item{d}{the long-memory parameter.}

\item{tau}{defines the search area, which is \code{[tau,1-tau]}. Default is \code{tau=0.15} as suggested by Andrews (1993).}
}
\value{
Returns the test statistic and the corresponding critical values of the test.
}
\description{
This function performs a Wilcoxon test for a change-in-mean that is robust under long memory. In contrast to a standard Wilcoxon test
it applies a self-normalization approach to estimate the long-run variance.
The function returns the test statistic as well as critical values.
}
\details{
Note that the critical values are generated for \code{tau=0.15}. Furthermore, it is assumed that we have a 1st-order Hermite process. For details see Betken (2016).
}
\examples{
library(fracdiff)
library(LongMemoryTS)

n        <- 500
d        <- 0.2
tseries  <- fracdiff.sim(n=n,d=d)$series
d_est    <- local.W(tseries, m=floor(1+n^0.65))$d

changep  <- c(rep(0,n/2),rep(1,n/2))
tseries2 <- tseries+changep
d_est2   <- local.W(tseries2, m=floor(1+n^0.65))$d

snwilcoxon(tseries,d=d_est)
snwilcoxon(tseries2,d=d_est2)
}
\references{
Wenger, K. and Leschinski, C. and Sibbertsen, P. (2018): Change-in-mean tests in long-memory time series: a review of recent developments. AStA Advances in Statistical Analysis, 103:2, pp. 237-256.

Betken, A. (2016): Testing for change-points in long-range dependent time series by means of a self-normalized wilcoxon test. Journal of Time Series Analysis, 37, pp. 785-908.

Andrews, D. W. K. (1993): Tests for Parameter Instability and Structural Change With Unknown Change Point. Econometrica, 61, pp. 821-856.
}
\author{
Kai Wenger
}
