% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_test.R
\name{wald_test}
\alias{wald_test}
\title{function to calculate wald_test. For internal use only}
\usage{
wald_test(Sigma, b, Terms)
}
\description{
function to calculate wald_test. For internal use only
}
\keyword{internal}
