% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBI_test.R
\name{LBI}
\alias{LBI}
\title{function to calculate sequence of LBI test statistics by Busetti and Taylor (2004). For internal use only.}
\usage{
LBI(x, trend, tau)
}
\description{
function to calculate sequence of LBI test statistics by Busetti and Taylor (2004). For internal use only.
}
\keyword{internal}
